ADC improvements, better units, valves added
Added valve controls
Pressures and pressure sensor temperatures are now reported as floats (bar and °C respectively)
ADC init forcibly enables channel 8 when temperature reading is enabled, to guarantee that temperature can actually be read
ADC init returns the wait time needed before first set of results is available
Added result error checking (and new error codes) to O2 conversions

+++ jamventProto/jamhwlib/I2cMS5607.h
@@ -11,8 +11,8 @@ public:
 		memset(_prom, 0, sizeof(_prom));
 	}
 
-	uint32_t  convertTemperature(uint32_t rawValue);
-	uint32_t convertPressure(uint32_t rawValue);
+	float convertTemperature(uint32_t rawValue);
+	float convertPressure(uint32_t rawValue);
 
 protected:
 	virtual  int getPrecisionDelay(unsigned precision);
